requirements_dev.txt
# Django核心
Django==5.2.4
djangorestframework==3.14.0
django-cors-headers==4.3.1

# 数据处理和分析
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2

# 图像处理
Pillow==10.0.1

# 开发工具
python-dotenv==1.0.0 